+++ Wagon.py
@@ -1,15 +1,21 @@
 class Wagon:
     """Primary administration object of a game
 
+    Class Variables:
+        stat_keys = List of strings - used to popular instance v. stats
+
     Instance Variables:
+        stats - dict - key strings, value int
         most = int
-        fear = int
-        hate = int
-        gold = int
-        rations = int
-        location = Tuple
-        pillagers = int
+
+    Instance Methods:
+        __init__ - make instance of Wagon
+        __str__ - represent Wagon instance in string form
+        change_stat - modify instance variable self.stats
+        check_loss - check if pillagers or rations are 0
     """
+
+    stat_keys = ["fear", "hate", "gold", "rations", "pillagers"]
 
     def __init__(self, most):
         """Constructor for class Wagon objects.
@@ -18,216 +24,55 @@ class Wagon:
             most = int
 
         Instance Variables:
-            fear = int
-            hate = int
-            gold = int
-            rations = int
-            location = tuple
-            pillagers = int
-            max = int
+            stats - dict - keys taken from class v. stat_keys
+            most = int
         """
-        self.fear = 0
-        self.hate = 0
-        self.gold = 0
-        self.rations = 0
-        self.location = (0, 0)
-        self.pillagers = 0
+        self.stats = {}
+        for stat in Wagon.stat_keys:
+            self.stats[stat] = 0
         self.most = most
 
     def __str__(self):
-        return f"""fear = {self.fear},
-                   hate = {self.hate},
-                   gold = {self.gold},
-                   rations = {self.rations},
-                   location = {self.location}
-                   pillagers = {self.pillagers}
-                   """
+        """Represent object in string form."""
+        return f"{self.stats}, max = {self.most}"
 
     def change_stat(self, stat, amount):
-        """Set method for external objects.
+        """modify matching value in self.stats by given amount.
 
         Parameters:
-        stat - must be 'fear', 'hate', 'gold', 'rations' or 'location
-        else ValueError
-        amount - int or tuple of ints for location
-        else TypeError
+        stat must be in self.stats else KeyError
+        amount int else ValueError
 
         Returns:
-        Boolean - of successful instance variable change made
+        int - new amount of relevant stat
 
         Dependencies:
-        self.change_gold
-        self.change_rations
-        self.change_location
-        self.change_raiders
-        self.change_hate
-        self.change_fear
+        self.stats
+        self.most
+        KeyError
         ValueError
-        self.max
          '"""
-        output = False
-        if stat == "fear":
-            output = self.change_fear(amount)
-        elif stat == "hate":
-            output = self.change_hate(amount)
-        elif stat == "gold":
-            output = self.change_gold(amount)
-        elif stat == "rations":
-            output = self.change_rations(amount)
-        elif stat == "location":
-            output = self.change_location(amount)
-        elif stat == "pillagers":
-            output = self.change_pillagers(amount)
+        output = 0
+        if not isinstance(amount, int):
+            raise ValueError
+        elif stat not in self.stats:
+            raise KeyError
         else:
-            raise ValueError("""Stat not recognised. Must be:
-                fear,
-                hate,
-                gold,
-                rations,
-                pillagers
-                or location""")
-        return output
-
-    def change_fear(self, amount):
-        """Modify fear instance variable.
-
-        Parameters:
-        amount - int
-        else TypeError
-
-        Returns:
-        boolean - if successfull modification made
-
-        Dependencies:
-        self.max
-        """
-        output = False
-        if self.fear + amount < 0:
-            self.fear = 0
-        elif self.fear + amount > self.most:
-            self.fear = self.most
-            output = True
-        else:
-            self.fear += amount
-            output = True
-        return output
-
-    def change_hate(self, amount):
-        """Modify hate instance variable.
-
-        Parameters:
-        amount - int
-        else TypeError
-
-        Returns:
-        boolean - if successfull modification made
-
-        Dependencies:
-        self.max
-        """
-        output = False
-        if self.hate + amount < 0:
-            self.hate = 0
-        elif self.hate + amount > self.most:
-            self.hate = self.most
-            output = True
-        else:
-            self.hate += amount
-            output = True
-        return output
-
-    def change_gold(self, amount):
-        """Modify gold instance variable.
-
-        Parameters:
-        amount - int
-        else TypeError
-
-        Returns:
-        boolean - if successfull modification made
-
-        Dependencies:
-        self.max
-        """
-        output = False
-        if self.gold + amount < 0:
-            self.gold = 0
-        elif self.gold + amount > self.most:
-            self.gold = self.most
-            output = True
-        else:
-            self.gold += amount
-            output = True
-        return output
-
-    def change_rations(self, amount):
-        """Modify rations instance variable.
-
-        Parameters:
-        amount - int
-        else TypeError
-
-        Returns:
-        boolean - if successfull modification made
-
-        Dependencies:
-        self.max
-        """
-        output = False
-        if self.rations + amount < 0:
-            self.rations = 0
-        elif self.rations + amount > self.most:
-            self.rations = self.most
-            output = True
-        else:
-            self.rations += amount
-            output = True
-        return output
-
-    def change_location(self, amount):
-        """Modify location instance variable.
-
-        Parameters:
-        amount - Tuple
-        else TypeError
-
-        Returns:
-        boolean - if successfull modification made
-
-        Dependencies:
-        self.max
-        """
-        pass
-
-    def change_pillagers(self, amount):
-        """Modify raiders instance variable.
-
-        Parameters:
-        amount - int
-        else TypeError
-
-        Returns:
-        boolean - if successfull modification made
-
-        Dependencies:
-        self.max
-        """
-        output = False
-        if self.pillagers + amount < 0:
-            self.pillagers = 0
-        elif self.pillagers + amount > self.most:
-            self.pillagers = self.most
-            output = True
-        else:
-            self.pillagers += amount
-            output = True
+            og_amount = self.stats[stat]
+            if og_amount + amount < 0:
+                output = 0
+            elif og_amount + amount > self.most:
+                output = self.most
+            else:
+                output = og_amount + amount
+        self.stats[stat] = output
         return output
 
     def check_loss(self):
         """Check if rations or pillagers are 0. True if so."""
         output = False
-        if self.rations == 0:
+        if self.stats["rations"] == 0:
             output = True
-        elif self.pillagers == 0:
+        elif self.stats["pillagers"] == 0:
             output = True
         return output